+++ app/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 
 [dependencies]
 lucien_render = { path = "../render" }
+lucien_core = { path = "../core" }
 
 iced = { version = "0.2.0", features = ["image", "tokio"] }
 iced_futures = "0.2"
