+++ src/main.rs
@@ -1,6 +1,4 @@
 use lucien_app as app;
-use lucien_core as core;
-use lucien_render as render;
 
 use anyhow::Result;
